fix: replace float16 casts with float32 and remove deprecated numpy aliases
- dataset.py/mcad.py: xarray DataArray.astype('float16') -> np.float32.
  New scipy/numpy no longer accepts 'float16' as an output dtype string
  ('Output dtype not compatible with inputs'), breaking count_matrix
  and mcad generation paths.
- utilities.py: np.issubdtype(np.object)/np.unicode are removed numpy
  aliases (numpy>=1.24); use builtin object/str instead.
- Preserves seekgene fork private changes (taps, cell_saturation metrics,
  cmeta fixes); dtype upcast float16->float32 is memory-neutral for
  typical bin-level counts and avoids scipy cast failures.

+++ ALLCools/count_matrix/dataset.py
@@ -31,7 +31,7 @@ def _bin_sf(cov, mc, p):
 def _cell_sf(cell_count_df):
     mc_sum, cov_sum = cell_count_df.sum()
     p = mc_sum / (cov_sum + 0.000001)  # prevent empty allc error
-    pv = cell_count_df.apply(lambda x: _bin_sf(x["cov"], x["mc"], p), axis=1).astype("float16")
+    pv = cell_count_df.apply(lambda x: _bin_sf(x["cov"], x["mc"], p), axis=1).astype(np.float32)
     return pv
 
 
@@ -203,7 +203,7 @@ def _calculate_pv(data, reverse_value, obs_dim, var_dim, cutoff=0.9):
     # get rid of small values, save space and memory
     pv[pv < cutoff] = 0
     pv = xr.DataArray(pv, coords=[data.coords[obs_dim], data.coords[var_dim]], dims=[obs_dim, var_dim])
-    pv = pv.astype("float16")
+    pv = pv.astype(np.float32)
     return pv
